Extract ChangePasswordView
